feat(requireAuth): Moving reusable logic to HOC
