+++ src/components/CommentBox.js
@@ -6,20 +6,6 @@ class CommentBox extends Component {
 
   state = { comment: '' };
 
-  componentDidMount() {
-    this.shouldNavigateAway();
-  }
-
-  componentDidUpdate(prevProps, prevState, snapshot) {
-    this.shouldNavigateAway();
-  }
-
-  shouldNavigateAway() {
-    if (!this.props.auth) {
-      this.props.history.push("/");
-    }
-  }
-
   handleChange = (e) => {
     this.setState({comment: e.target.value});
   };
@@ -48,8 +34,4 @@ class CommentBox extends Component {
   }
 }
 
-function mapStateToProps(state) {
-  return { auth: state.auth };
-}
-
-export default connect(mapStateToProps, actions)(CommentBox);
+export default connect(null, actions)(CommentBox);
