+++ src/components/requireAuth.js
@@ -1,11 +1,31 @@
 import React, { Component } from "react";
+import { connect } from "react-redux";
 
 export default (ChildComponent) => {
   class ComposedComponent extends Component {
+
+    componentDidMount() {
+      this.shouldNavigateAway();
+    }
+
+    componentDidUpdate(prevProps, prevState, snapshot) {
+      this.shouldNavigateAway();
+    }
+
+    shouldNavigateAway() {
+      if (!this.props.auth) {
+        this.props.history.push("/");
+      }
+    }
+
     render() {
       return <ChildComponent/>;
     }
   }
 
-  return ComposedComponent;
+  function mapStateToProps(state) {
+    return { auth: state.auth };
+  }
+
+  return connect(mapStateToProps)(ComposedComponent);
 };
